backend/requirements.txt
# GoBus Backend - Django REST Framework

Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.2.0
qrcode==7.4.2
razorpay==1.4.1
firebase-admin==6.3.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
